requirements.txt
playwright>=1.40.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
functions-framework>=3.5.0
